Started Extraction Of SNK 40th Anniversary Collection
I noticed that the upstream Master had merged in a change by @ghoost82 and I thought, as the SNK 40th Anniversary Collection also uses Digital Extreme's MOO emulator, it'd be a good idea to add it to the extraction functionality too.

There are other tools that already do this and once again this is just the work of ghoost82 reimplemented.

Currently only converts Beast Busters.