Started work on ssf2tu
Started working on Super Street Fighter 2 Turbo but the graphics have been scrambled in a way I can't figure out.  Based on how the MAME project works, it should be twelve files split into three sets of four, which are then interleaved together.  But I've done that and it's still wrong.  So there must be something else going on.